+++ client/src/App.jsx
@@ -2,12 +2,16 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import Login from './pages/Login';
 import Home from './pages/Home';
+import AlumniInvite from './pages/AlumniInvite';
+import CompleteProfile from './pages/CompleteProfile';
 
 function App() {
   return (
     <Router>
       <Routes>
         <Route path="/login" element={<Login />} />
+        <Route path="/alumni-invite/:token" element={<AlumniInvite />} />
+        <Route path="/complete-profile" element={<CompleteProfile />} />
         <Route path="/home" element={<Home />} />
         <Route path="/" element={<Navigate to="/home" replace />} />
       </Routes>